requirements.txt
langchain>=1.0
langgraph>=1.0
langchain-openai
langchain-core
langchain-mcp-adapters
agentevals
chainlit
python-dotenv
pydantic>=2.0
